Fix bug on ppc64 caused by narrowing cast (#941)
The narrowing cast happens with the second argument in the call to
u_set_char() in format_parse() in format_print.c. The argument is
str->len.

Before this commit, &str->len was cast to (int *), despite str being
a "struct gbuf" whose len member has type size_t.

This does not cause an issue on 64-bit little-endian systems because a
cast to (int *), though narrowing, points to the bottom four bytes of
str->len, and so the incrementing of the local copy of the idx
parameter in u_set_char() and then the assignment back to *idx doesn't
have any undesired behavior.

64-bit little-endian
|0|1|2|3|4|5|6|7| (str->len, numbers indicate byte significance)
 ^
 |
 |_(size_t *)

|0|1|2|3|
 ^
 |_(int *)

An increment here on 64-bit little-endian increases the value held in
byte 0, which is the expected result, so narrowing has no detrimental
effect.

However the narrowing cast causes a bug on big-endian 64-bit
systems (such as ppc64) because a cast to (int *) points to the top
four bytes of str->len.

64-bit big-endian
|7|6|5|4|3|2|1|0| (str->len)
 ^
 |
 |_(size_t *)

|7|6|5|4|
 ^
 |_(int *)

Besides u_set_char() being passed the top 4 bytes when the bottom 4
bytes are desired, an increment of the local copy of idx in
u_set_char() and the assignment back to *idx increases the value held
in byte 4, when byte 0 should be incremented. So the assignment back
to *idx increases the value of str->len by 2^32, and thus the next
time str->len is used for array access a segfault occurs.

This commit fixes the bug by changing the idx parameter of
u_set_char() from type (int *) to type (size_t *), and removing the
cast to (int *) in format_parse().

This commit also updates all functions that use u_set_char() to pass
a (size_t *) argument in order to eliminate any warnings about
incompatible pointer types.

+++ id3.c
@@ -287,10 +287,10 @@ static int utf16_is_special(uchar uch)
 	return utf16_is_hsurrogate(uch) || utf16_is_lsurrogate(uch) || utf16_is_bom(uch);
 }
 
-static char *utf16_to_utf8(const unsigned char *buf, int buf_size)
+static char *utf16_to_utf8(const unsigned char *buf, size_t buf_size)
 {
 	char *out;
-	int i, idx;
+	size_t i, idx;
 	int little_endian = 0;
 
 	if (buf_size < 2)
